consolidate the CheckForInterrupts*Nodes
